refactor: clean up cloud settings and analytics
