+++ src/app.tsx
@@ -20,7 +20,7 @@ import { usePageTracking } from '@/hooks/use-analytics'
 import { useKeyboardInset } from '@/hooks/use-keyboard-inset'
 import { useMcpSync } from '@/hooks/use-mcp-sync'
 import ChatLayout from '@/layout/main-layout'
-import { initPosthog, PostHogProvider } from '@/lib/analytics'
+import { initPosthog, PostHogProvider } from '@/lib/posthog'
 import { reconcileDefaults } from '@/lib/reconcile-defaults'
 import { ThemeProvider } from '@/lib/theme-provider'
 import DevSettingsPage from '@/settings/dev-settings'
