+++ src/automations/index.tsx
@@ -21,7 +21,7 @@ import { triggersTable } from '@/db/tables'
 import { defaultAutomations } from '@/defaults/automations'
 import { isAutomationModified } from '@/defaults/utils'
 import { useSettings } from '@/hooks/use-settings'
-import { trackEvent } from '@/lib/analytics'
+import { trackEvent } from '@/lib/posthog'
 import { cn } from '@/lib/utils'
 import type { Prompt, Trigger } from '@/types'
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query'
