+++ src/chats/chat-state.tsx
@@ -2,7 +2,7 @@ import { aiFetchStreamingResponse } from '@/ai/fetch'
 import ChatUI from '@/components/chat/chat-ui'
 import { useSettings } from '@/hooks/use-settings'
 import { useThrottledCallback } from '@/hooks/use-throttle'
-import { trackEvent } from '@/lib/analytics'
+import { trackEvent } from '@/lib/posthog'
 import { getDefaultModelForThread, getTriggerPromptForThread } from '@/dal'
 import { useMCP } from '@/lib/mcp-provider'
 import type { Model, SaveMessagesFunction, ThunderboltUIMessage } from '@/types'
