+++ src/components/chat/tool-icon.tsx
@@ -46,7 +46,7 @@ export const getProxiedFaviconUrl = (faviconUrl: string, cloudUrl: string): stri
  */
 const useToolFavicon = (toolName: string, toolOutput: unknown, isLoading: boolean, isError: boolean) => {
   const [failedFavicons, setFailedFavicons] = useState<Set<string>>(new Set())
-  const { cloudUrl } = useSettings({ cloud_url: String })
+  const { cloudUrl } = useSettings({ cloud_url: 'http://localhost:8000/v1' })
 
   const handleFaviconError = (url: string) => {
     setFailedFavicons((prev) => new Set(prev).add(url))
