+++ src/hooks/use-units-options.ts
@@ -1,8 +1,8 @@
+import { getSettings } from '@/dal'
+import { unitsOptionsResponseSchema } from '@/schemas/api'
+import type { UnitsOptionsData } from '@/types'
 import { useQuery } from '@tanstack/react-query'
 import ky from 'ky'
-import type { UnitsOptionsData } from '@/types'
-import { unitsOptionsResponseSchema } from '@/schemas/api'
-import { getCloudUrl } from '@/lib/config'
 
 /**
  * Fetches units options data from the backend API
@@ -12,7 +12,7 @@ export const useUnitsOptions = () => {
   return useQuery({
     queryKey: ['units-options'],
     queryFn: async (): Promise<UnitsOptionsData> => {
-      const cloudUrl = await getCloudUrl()
+      const { cloudUrl } = await getSettings({ cloud_url: 'http://localhost:8000/v1' })
       const response = await ky.get(`${cloudUrl}/units-options`).json()
 
       const validatedData = unitsOptionsResponseSchema.parse(response)
