+++ src/integrations/microsoft/auth.ts
@@ -1,13 +1,13 @@
+import { getSettings } from '@/dal'
 import type { GoogleUserInfo } from '@/integrations/google/types'
 import type { OAuthConfig, OAuthTokens } from '@/lib/auth'
-import { getCloudUrl } from '@/lib/config'
 import { memoize } from '@/lib/memoize'
 import { isTauri } from '@/lib/platform'
 import type { AuthProviderBackendConfig } from '@/types'
 import ky from 'ky'
 
 const fetchBackendConfig = memoize(async (): Promise<AuthProviderBackendConfig> => {
-  const cloudUrl = await getCloudUrl()
+  const { cloudUrl } = await getSettings({ cloud_url: 'http://localhost:8000/v1' })
   return await ky.get(`${cloudUrl}/auth/microsoft/config`).json<AuthProviderBackendConfig>()
 })
 
@@ -38,7 +38,7 @@ export const buildAuthUrl = async (state: string, codeChallenge: string): Promis
 
 export const exchangeCodeForTokens = async (code: string, codeVerifier: string): Promise<OAuthTokens> => {
   const config = await getOAuthConfig()
-  const cloudUrl = await getCloudUrl()
+  const { cloudUrl } = await getSettings({ cloud_url: 'http://localhost:8000/v1' })
   return await ky
     .post(`${cloudUrl}/auth/microsoft/exchange`, {
       json: { code, code_verifier: codeVerifier, redirect_uri: config.redirectUri },
@@ -63,7 +63,7 @@ export const getUserInfo = async (accessToken: string): Promise<GoogleUserInfo> 
 }
 
 export const refreshAccessToken = async (refreshToken: string): Promise<OAuthTokens> => {
-  const cloudUrl = await getCloudUrl()
+  const { cloudUrl } = await getSettings({ cloud_url: 'http://localhost:8000/v1' })
   return await ky
     .post(`${cloudUrl}/auth/microsoft/refresh`, { json: { refresh_token: refreshToken } })
     .json<OAuthTokens>()
