+++ src/lib/http-client.ts
@@ -1,5 +1,5 @@
+import { getSettings } from '@/dal'
 import ky from 'ky'
-import { getCloudUrl } from './config'
 
 // Cache for the HTTP client as a singleton instance to avoid recreating it
 let singleton: typeof ky | null = null
@@ -10,7 +10,7 @@ export const getHttpClient = async (): Promise<typeof ky> => {
     return singleton
   }
 
-  const cloudUrl = await getCloudUrl()
+  const { cloudUrl } = await getSettings({ cloud_url: 'http://localhost:8000/v1' })
 
   singleton = ky.create({
     prefixUrl: cloudUrl,
