+++ src/tasks/index.tsx
@@ -2,10 +2,10 @@ import { Button } from '@/components/ui/button'
 import { SearchInput } from '@/components/ui/search-input'
 import { Skeleton } from '@/components/ui/skeleton'
 import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from '@/components/ui/tooltip'
+import { getIncompleteTasks, getIncompleteTasksCount, updateTask } from '@/dal'
 import { DatabaseSingleton } from '@/db/singleton'
 import { tasksTable } from '@/db/tables'
-import { trackEvent } from '@/lib/analytics'
-import { getIncompleteTasks, getIncompleteTasksCount, updateTask } from '@/dal'
+import { trackEvent } from '@/lib/posthog'
 import { cn } from '@/lib/utils'
 import type { Task } from '@/types'
 import type { DropAnimation } from '@dnd-kit/core'
